Defaults and Precision Settings

qAlgebra allows configuration of the number formatting used for coefficients, including float and exponential precision. These defaults are persistent across sessions using Preferences.jl.

API

qAlgebra.get_default — Function

get_default(name::Symbol)

Returns the current default value for the coefficient preference with the given name. If no such preference has been set it returns the default value from DEFAULT_COEFF_PREFS.

qAlgebra.set_flip_if_first_term_negative — Function

set_flip_if_first_term_negative(mode::Bool)

Sets a new default value for the first mode and saves it persistently. First mode specifies whether braced terms with a leading negative are flipped or only if all terms are negative.

qAlgebra.set_do_braced — Function

set_do_braced(b::Bool)

Sets a new default value for :DO_BRACED. Toggles whether terms are grouped when printing them, into groups with common coefficients.